+++ source/CMakeLists.txt
@@ -365,7 +365,9 @@ target_sources(EndlessSkyLib PRIVATE
 	comparators/ByName.h
 	comparators/BySeriesAndIndex.h
 	concurrent/Parallel.h
-	concurrent/PartiallyGuarded.h
+	concurrent/PartiallyGuardedList.h
+	concurrent/PartiallyGuardedMap.h
+	concurrent/PartiallyGuardedVector.h
 	concurrent/ResourceProvider.h
 	concurrent/WithMutex.cpp
 	concurrent/WithMutex.h
